Adding group functionality to dynamo backend
